fix: Cast BitMath return values to uint256 for proper comparison

+++ contracts/milestone2/test/BitMathTest.sol
@@ -17,26 +17,26 @@ contract BitMathTest is Test {
     }
 
     function testMostSignificantBitOne() public {
-        assertEq(BitMath.mostSignificantBit(1), 0);
+        assertEq(uint256(BitMath.mostSignificantBit(1)), uint256(0));
     }
 
     function testLeastSignificantBitOne() public {
-        assertEq(BitMath.leastSignificantBit(1), 0);
+        assertEq(uint256(BitMath.leastSignificantBit(1)), uint256(0));
     }
 
     function testMostSignificantBitPowersOfTwo() public {
-        assertEq(BitMath.mostSignificantBit(2), 1);
-        assertEq(BitMath.mostSignificantBit(4), 2);
-        assertEq(BitMath.mostSignificantBit(8), 3);
-        assertEq(BitMath.mostSignificantBit(16), 4);
-        assertEq(BitMath.mostSignificantBit(32), 5);
+        assertEq(uint256(BitMath.mostSignificantBit(2)), uint256(1));
+        assertEq(uint256(BitMath.mostSignificantBit(4)), uint256(2));
+        assertEq(uint256(BitMath.mostSignificantBit(8)), uint256(3));
+        assertEq(uint256(BitMath.mostSignificantBit(16)), uint256(4));
+        assertEq(uint256(BitMath.mostSignificantBit(32)), uint256(5));
     }
 
     function testLeastSignificantBitPowersOfTwo() public {
-        assertEq(BitMath.leastSignificantBit(2), 1);
-        assertEq(BitMath.leastSignificantBit(4), 2);
-        assertEq(BitMath.leastSignificantBit(8), 3);
-        assertEq(BitMath.leastSignificantBit(16), 4);
-        assertEq(BitMath.leastSignificantBit(32), 5);
+        assertEq(uint256(BitMath.leastSignificantBit(2)), uint256(1));
+        assertEq(uint256(BitMath.leastSignificantBit(4)), uint256(2));
+        assertEq(uint256(BitMath.leastSignificantBit(8)), uint256(3));
+        assertEq(uint256(BitMath.leastSignificantBit(16)), uint256(4));
+        assertEq(uint256(BitMath.leastSignificantBit(32)), uint256(5));
     }
 }
